redesign match

+++ tipovacka-app/src/app/components/match/match.component.ts
@@ -1,4 +1,4 @@
-import {Component, inject, Input} from '@angular/core';
+import {Component, inject, Input, OnChanges, OnInit} from '@angular/core';
 import {MatchWithTeamName} from '../../models/match.model';
 import {AuthService} from '../../services/auth.service';
 import {Vote, VoteResult} from '../../models/vote.model';
@@ -9,6 +9,14 @@ import {ResultButtonComponent} from '../result-button/result-button.component';
 import {first} from 'rxjs';
 import {DialogService} from '@ngneat/dialog';
 import {SignInAlertComponent} from '../sign-in-alert/sign-in-alert.component';
+import {DatetimeFormatPipe} from '../../pipes/datetime-format.pipe';
+import {ImageSrcErrorDirective} from '../../directives/imgSrcError.directive';
+
+type TeamState =
+  'normal'
+  | 'highlight'
+  | 'correct'
+  | 'incorrect';
 
 @Component({
   selector: 'app-match',
@@ -20,17 +28,30 @@ import {SignInAlertComponent} from '../sign-in-alert/sign-in-alert.component';
     TranslocoPipe,
     AsyncPipe,
     ResultButtonComponent,
-    NgClass
+    NgClass,
+    DatetimeFormatPipe,
+    ImageSrcErrorDirective
   ]
 })
-export class MatchComponent {
+export class MatchComponent implements OnChanges, OnInit {
   @Input({required: true}) data!: MatchWithTeamName;
-  @Input() votes: Record<string, Vote | undefined> = {};
+  @Input() votes?: Record<string, Vote | undefined>;
 
   now = new Date();
 
-  sendingVote: VoteResult | undefined = undefined;
+  sendingVote?: VoteResult = undefined;
   public authService = inject(AuthService);
+  states: Record<1 | 0 | 2, TeamState> = {
+    1: 'normal',
+    0: 'normal',
+    2: 'normal',
+  }
+  hides: Record<1 | 0 | 2, boolean> = {
+    1: false,
+    0: false,
+    2: false,
+  }
+  protected readonly undefined = undefined;
   private dataService = inject(DataService);
   private dialogService = inject(DialogService);
 
@@ -46,7 +67,7 @@ export class MatchComponent {
             return;
           }
           this.sendingVote = chosenResult;
-          if (chosenResult === this.votes[this.data.id]?.result) {
+          if (this.votes && chosenResult === this.votes[this.data.id]?.result) {
             this.dataService.deleteVote(this.data.id)
               .then(
                 () => {
@@ -63,5 +84,39 @@ export class MatchComponent {
             )
         }
       })
+  }
+
+  ngOnInit() {
+    this.calculateStates();
+  }
+
+  ngOnChanges() {
+    this.calculateStates();
+  }
+
+  calculateStates = () => {
+    this.states['1'] = 'normal';
+    this.states['0'] = 'normal';
+    this.states['2'] = 'normal';
+    if (this.data.datetime > this.now) {
+      if (!this.votes || !this.votes[this.data.id] || this.votes[this.data.id]?.result === null) {
+        return;
+      }
+      this.hides['1'] = true;
+      this.hides['0'] = true;
+      this.hides['2'] = true;
+      this.states[this.votes[this.data.id]!.result] = 'highlight';
+      this.hides[this.votes[this.data.id]!.result] = false;
+      return;
+    }
+
+    this.hides['1'] = true;
+    this.hides['0'] = true;
+    this.hides['2'] = true;
+    if (this.votes && this.votes[this.data.id]) {
+      this.states[this.votes[this.data.id]!.result] = 'incorrect';
+      this.hides[this.votes[this.data.id]!.result] = false;
+    }
+    this.states[this.data.result!] = 'correct';
   }
 }
